+++ enclave-manager/web/src/client/enclaveManager/KurtosisClient.ts
@@ -26,6 +26,7 @@ import {
 import { EnclaveFullInfo } from "../../emui/enclaves/types";
 import { assertDefined, asyncResult, isDefined } from "../../utils";
 import { RemoveFunctions } from "../../utils/types";
+import { KURTOSIS_CLOUD_HOST } from "../constants";
 
 export abstract class KurtosisClient {
   protected readonly client: PromiseClient<typeof KurtosisEnclaveManagerServer>;
@@ -50,8 +51,23 @@ export abstract class KurtosisClient {
     this.client = client;
     this.cloudUrl = parentUrl;
     this.baseApplicationUrl = childUrl;
-    console.log("cloudUrl", this.cloudUrl);
-    console.log("baseApplicationUrl", this.baseApplicationUrl);
+    this.getParentRequestedRoute();
+  }
+
+  getParentRequestedRoute() {
+    const splits = this.cloudUrl.pathname.split("/enclave-manager");
+    if (splits[1]) {
+      return splits[1];
+    }
+    return undefined;
+  }
+
+  getCloudUrl() {
+    return this.cloudUrl;
+  }
+
+  isRunningInCloud() {
+    return this.cloudUrl.host.toLowerCase().includes(KURTOSIS_CLOUD_HOST);
   }
 
   abstract getHeaderOptions(): { headers?: Headers };