+++ enclave-manager/web/src/components/CodeEditor.tsx
@@ -16,6 +16,8 @@ export const CodeEditor = ({ text, onTextChange, showLineNumbers }: CodeEditorPr
 
   const resizeEditorBasedOnContent = () => {
     if (isDefined(editor)) {
+      // An initial layout call is needed, else getContentHeight is garbage
+      editor.layout();
       const contentHeight = editor.getContentHeight();
       editor.layout({ width: 500, height: contentHeight });
       // Unclear why layout must be called twice, but seems to be necessary
@@ -46,12 +48,13 @@ export const CodeEditor = ({ text, onTextChange, showLineNumbers }: CodeEditorPr
   resizeEditorBasedOnContent();
 
   return (
-    <Box width={"100%"}>
+    <Box width={"100%"} maxHeight={"1000px"}>
       <Editor
         onMount={handleMount}
         value={text}
         onChange={handleChange}
         options={{
+          automaticLayout: false, // if this is `true` a ResizeObserver is installed. This causes issues with us managing the container size outside.
           readOnly: isReadOnly,
           lineNumbers: showLineNumbers || (!isDefined(showLineNumbers) && !isReadOnly) ? "on" : "off",
           minimap: { enabled: false },
@@ -68,7 +71,6 @@ export const CodeEditor = ({ text, onTextChange, showLineNumbers }: CodeEditorPr
         }}
         defaultLanguage={"json"}
         theme={"vs-dark"}
-        height={"100%"}
       />
     </Box>
   );