+++ enclave-manager/web/src/components/enclaves/modals/ConfigureEnclaveModal.tsx
@@ -11,13 +11,15 @@ import {
   ModalOverlay,
   Text,
   Tooltip,
+  useToast,
 } from "@chakra-ui/react";
 import { EnclaveMode } from "enclave-manager-sdk/build/engine_service_pb";
 import { useMemo, useRef, useState } from "react";
 import { SubmitHandler } from "react-hook-form";
-import { useNavigate, useSubmit } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 import { useKurtosisClient } from "../../../client/enclaveManager/KurtosisClientContext";
 import { ArgumentValueType, KurtosisPackage } from "../../../client/packageIndexer/api/kurtosis_package_indexer_pb";
+import { useEmuiAppContext } from "../../../emui/EmuiAppContext";
 import { EnclaveFullInfo } from "../../../emui/enclaves/types";
 import { assertDefined, isDefined, stringifyError } from "../../../utils";
 import { KURTOSIS_PACKAGE_ID_URL_ARG, KURTOSIS_PACKAGE_PARAMS_URL_ARG } from "../../constants";
@@ -49,14 +51,15 @@ export const ConfigureEnclaveModal = ({
   existingEnclave,
 }: ConfigureEnclaveModalProps) => {
   const kurtosisClient = useKurtosisClient();
+  const { createEnclave, runStarlarkPackage } = useEmuiAppContext();
   const navigator = useNavigate();
-  const submit = useSubmit();
   const [isLoading, setIsLoading] = useState(false);
   const [error, setError] = useState<string>();
   const formRef = useRef<EnclaveConfigurationFormImperativeAttributes>(null);
+  const toast = useToast();
 
   const initialValues = useMemo(() => {
-    if (isDefined(existingEnclave)) {
+    if (isDefined(existingEnclave) && isDefined(existingEnclave.starlarkRun)) {
       if (existingEnclave.starlarkRun.isErr) {
         setError(
           `Could not retrieve starlark run for previous configuration, got error: ${existingEnclave.starlarkRun.isErr}`,
@@ -73,13 +76,11 @@ export const ConfigureEnclaveModal = ({
         ): any => {
           switch (argType) {
             case ArgumentValueType.BOOL:
-              return !!value ? "true" : "false";
+              return !!value ? "true" : isDefined(value) ? "false" : "";
             case ArgumentValueType.INTEGER:
               return isDefined(value) ? `${value}` : "";
             case ArgumentValueType.STRING:
               return value || "";
-            case ArgumentValueType.JSON:
-              return isDefined(value) ? JSON.stringify(value) : "{}";
             case ArgumentValueType.LIST:
               assertDefined(innerType1, `Cannot parse a list argument type without knowing innerType1`);
               return isDefined(value) ? value.map((v: any) => convertArgValue(innerType1, v)) : [];
@@ -88,8 +89,10 @@ export const ConfigureEnclaveModal = ({
               return isDefined(value)
                 ? Object.entries(value).map(([k, v]) => ({ key: k, value: convertArgValue(innerType2, v) }), {})
                 : [];
+            case ArgumentValueType.JSON:
             default:
-              return value;
+              // By default, a typeless parameter is JSON.
+              return isDefined(value) ? JSON.stringify(value) : "{}";
           }
         };
 
@@ -120,7 +123,20 @@ export const ConfigureEnclaveModal = ({
     if (!isDefined(preloadArgs)) {
       return undefined;
     }
-    return JSON.parse(atob(preloadArgs)) as ConfigureEnclaveForm;
+    const parsedForm = JSON.parse(atob(preloadArgs)) as ConfigureEnclaveForm;
+    kurtosisPackage.args
+      .filter((arg) => !isDefined(arg.typeV2?.topLevelType) || arg.typeV2?.topLevelType === ArgumentValueType.JSON)
+      .forEach((arg) => {
+        if (parsedForm.args[arg.name]) {
+          try {
+            parsedForm.args[arg.name] = JSON.stringify(JSON.parse(parsedForm.args[arg.name]), undefined, 4);
+          } catch (err: any) {
+            console.error("err", err);
+            // do nothing, the input was not valid json.
+          }
+        }
+      });
+    return parsedForm;
   }, [existingEnclave, kurtosisPackage.args]);
 
   const getLinkToCurrentConfig = () => {
@@ -133,18 +149,37 @@ export const ConfigureEnclaveModal = ({
   };
 
   const handleClose = () => {
-    navigator("#", { replace: true });
-    onClose();
+    if (!isLoading) {
+      navigator("#", { replace: true });
+      setError(undefined);
+      onClose();
+    }
   };
 
   const handleLoadSubmit: SubmitHandler<ConfigureEnclaveForm> = async (formData) => {
     setError(undefined);
 
+    try {
+      console.debug("formData", formData);
+      if (formData.args && formData.args.args) {
+        formData.args.args = JSON.parse(formData.args.args);
+        console.debug("successfully parsed args as proper JSON", formData.args.args);
+      }
+    } catch (err) {
+      toast({
+        title: `An error occurred while preparing data for running package. The package arguments were not proper JSON: ${stringifyError(
+          err,
+        )}`,
+        colorScheme: "red",
+      });
+      return;
+    }
+
     let apicInfo = existingEnclave?.apiContainerInfo;
     let enclaveUUID = existingEnclave?.shortenedUuid;
     if (!isDefined(existingEnclave)) {
       setIsLoading(true);
-      const newEnclave = await kurtosisClient.createEnclave(formData.enclaveName, "info", formData.restartServices);
+      const newEnclave = await createEnclave(formData.enclaveName, "info", formData.restartServices);
       setIsLoading(false);
 
       if (newEnclave.isErr) {
@@ -163,36 +198,72 @@ export const ConfigureEnclaveModal = ({
       setError(`Cannot trigger starlark run as apic info cannot be found`);
       return;
     }
-    submit(
-      { config: formData, packageId: kurtosisPackage.name, apicInfo: apicInfo.toJson() },
-      {
-        method: "post",
-        action: `/enclave/${enclaveUUID}/logs`,
-        encType: "application/json",
-      },
-    );
+
+    let submissionData = {};
+    if (formData.args.args) {
+      const { args, ...rest } = formData.args;
+      submissionData = {
+        ...args,
+        ...rest,
+      };
+      console.debug("formData has `args` field and is merged with other potential args", submissionData);
+    } else {
+      submissionData = {
+        ...formData.args,
+      };
+      console.debug("formData does not have Args field", submissionData);
+    }
+    console.log("submissionData for runStarlarkPackage", submissionData);
+
+    const logsIterator = await runStarlarkPackage(apicInfo, kurtosisPackage.name, submissionData);
+    navigator(`/enclave/${enclaveUUID}/logs`, { state: { logs: logsIterator } });
     onClose();
   };
 
   return (
-    <Modal closeOnOverlayClick={false} isOpen={isOpen} onClose={handleClose} isCentered size={"5xl"}>
+    <Modal
+      closeOnOverlayClick={false}
+      isOpen={isOpen}
+      onClose={handleClose}
+      isCentered
+      size={"5xl"}
+      scrollBehavior={"inside"}
+    >
       <ModalOverlay />
       <ModalContent>
-        <ModalHeader textAlign={"center"}>{!isDefined(existingEnclave) && "New "}Enclave Configuration</ModalHeader>
+        <ModalHeader flex={"0"} textAlign={"center"}>
+          {!isDefined(existingEnclave) && "New "}Enclave Configuration
+        </ModalHeader>
         <ModalCloseButton />
         <EnclaveConfigurationForm
           ref={formRef}
           initialValues={initialValues}
           onSubmit={handleLoadSubmit}
           kurtosisPackage={kurtosisPackage}
+          style={{
+            display: "flex",
+            flexDirection: "column",
+            flex: "0 1 auto",
+            minHeight: 0,
+          }}
         >
-          <ModalBody p={"0px"}>
-            <Flex fontSize={"sm"} justifyContent={"center"} alignItems={"center"} gap={"12px"} pb={"12px"}>
+          <ModalBody flex="0 1 auto" p={"0px"} display={"flex"} flexDirection={"column"}>
+            <Flex flex={"0"} fontSize={"sm"} justifyContent={"center"} alignItems={"center"} gap={"12px"} pb={"12px"}>
               <Text>Configuring</Text>
               <EnclaveSourceButton source={kurtosisPackage.name} size={"sm"} variant={"outline"} color={"gray.100"} />
             </Flex>
-            {isDefined(error) && <KurtosisAlert message={error} />}
-            <Flex flexDirection={"column"} gap={"24px"} p={"12px 24px"} bg={"gray.900"}>
+            {isDefined(error) && (
+              <KurtosisAlert flex={"1 0 auto"} message={"Could not execute configuration"} details={error} />
+            )}
+            <Flex
+              flex={"0 1 auto"}
+              overflowY={"scroll"}
+              minHeight={0}
+              flexDirection={"column"}
+              gap={"24px"}
+              p={"12px 24px"}
+              bg={"gray.900"}
+            >
               <Flex justifyContent={"space-between"} alignItems={"center"}>
                 <Tooltip
                   shouldWrapChildren
@@ -212,7 +283,7 @@ export const ConfigureEnclaveModal = ({
                   name={"enclaveName"}
                   disabled={isDefined(existingEnclave)}
                   validate={(value) => {
-                    if (!isEnclaveNameAllowed(value)) {
+                    if (value.length > 0 && !isEnclaveNameAllowed(value)) {
                       return `The enclave name must match ${allowedEnclaveNamePattern}`;
                     }
                   }}
@@ -223,9 +294,9 @@ export const ConfigureEnclaveModal = ({
               ))}
             </Flex>
           </ModalBody>
-          <ModalFooter>
+          <ModalFooter flex={"0"}>
             <Flex justifyContent={"flex-end"} gap={"12px"}>
-              <Button color={"gray.100"} onClick={handleClose} disabled={isLoading}>
+              <Button color={"gray.100"} onClick={handleClose} isDisabled={isLoading}>
                 Cancel
               </Button>
               <Button type={"submit"} isLoading={isLoading} colorScheme={"kurtosisGreen"}>