+++ enclave-manager/web/src/components/enclaves/tables/PortsTable.tsx
@@ -1,7 +1,10 @@
-import { Flex, Text } from "@chakra-ui/react";
+import { ExternalLinkIcon } from "@chakra-ui/icons";
+import { Flex, Icon, Link, Text, Tooltip } from "@chakra-ui/react";
 import { ColumnDef, createColumnHelper } from "@tanstack/react-table";
 import { Port } from "enclave-manager-sdk/build/api_container_service_pb";
 import { useMemo } from "react";
+import { FiAlertTriangle } from "react-icons/fi";
+import { useKurtosisClient } from "../../../client/enclaveManager/KurtosisClientContext";
 import { CopyButton } from "../../CopyButton";
 import { DataTable } from "../../DataTable";
 import { transportProtocolToString } from "../utils";
@@ -11,10 +14,16 @@ type PortsTableRow = {
   link: string;
 };
 
-const getPortTableRows = (privatePorts: Record<string, Port>, publicPorts: Record<string, Port>): PortsTableRow[] => {
+const getPortTableRows = (
+  privatePorts: Record<string, Port>,
+  publicPorts: Record<string, Port>,
+  publicIp: string,
+): PortsTableRow[] => {
   return Object.entries(privatePorts).map(([name, port]) => ({
     port: { transportProtocol: transportProtocolToString(port.transportProtocol), privatePort: port.number, name },
-    link: "Coming soon",
+    link: `${port.maybeApplicationProtocol ? port.maybeApplicationProtocol + "://" : ""}${publicIp}:${
+      publicPorts[name].number
+    }`,
   }));
 };
 
@@ -23,9 +32,12 @@ const columnHelper = createColumnHelper<PortsTableRow>();
 type PortsTableProps = {
   privatePorts: Record<string, Port>;
   publicPorts: Record<string, Port>;
+  publicIp: string;
 };
 
-export const PortsTable = ({ privatePorts, publicPorts }: PortsTableProps) => {
+export const PortsTable = ({ privatePorts, publicPorts, publicIp }: PortsTableProps) => {
+  const kurtosisClient = useKurtosisClient();
+
   const columns = useMemo<ColumnDef<PortsTableRow, any>[]>(
     () => [
       columnHelper.accessor("port", {
@@ -42,7 +54,28 @@ export const PortsTable = ({ privatePorts, publicPorts }: PortsTableProps) => {
       columnHelper.accessor("link", {
         header: "Link",
         minSize: 800,
-        cell: ({ row }) => <Text width={"100%"}>{row.original.link}</Text>,
+        cell: ({ row }) => (
+          <Text width={"100%"}>
+            {row.original.link.startsWith("http") ? (
+              <Link href={row.original.link} isExternal>
+                {row.original.link}
+                <ExternalLinkIcon mx="2px" />
+              </Link>
+            ) : (
+              row.original.link
+            )}
+            {kurtosisClient.isRunningInCloud() && (
+              <Tooltip
+                label={
+                  "Only enclaves started using the CLI will have their ports available. This port may not work if it was started using the app."
+                }
+                shouldWrapChildren
+              >
+                <Icon m="0 10px" as={FiAlertTriangle} color={"orange.400"} />
+              </Tooltip>
+            )}
+          </Text>
+        ),
       }),
       columnHelper.display({
         id: "copyButton",
@@ -52,19 +85,19 @@ export const PortsTable = ({ privatePorts, publicPorts }: PortsTableProps) => {
               contentName={"link"}
               isIconButton
               aria-label={"Copy this port"}
-              valueToCopy={`${row.original.port.privatePort}`}
+              valueToCopy={`${row.original.link}`}
             />
           </Flex>
         ),
       }),
     ],
-    [],
+    [kurtosisClient],
   );
 
   return (
     <DataTable
       columns={columns}
-      data={getPortTableRows(privatePorts, publicPorts)}
+      data={getPortTableRows(privatePorts, publicPorts, publicIp)}
       defaultSorting={[{ id: "number", desc: true }]}
     />
   );