+++ enclave-manager/web/src/emui/Navbar.tsx
@@ -8,6 +8,7 @@ export type NavbarProps = {
 
 export const Navbar = ({ baseApplicationUrl }: NavbarProps) => {
   const location = useLocation();
+  // const kurtosisClient = useKurtosisClient();
 
   return (
     <Navigation baseApplicationUrl={baseApplicationUrl}>
@@ -18,6 +19,11 @@ export const Navbar = ({ baseApplicationUrl }: NavbarProps) => {
           isActive={location.pathname === "/" || location.pathname.startsWith("/enclave")}
         />
       </Link>
+      {/*{kurtosisClient.isRunningInCloud() && (*/}
+      {/*  <Link to={KURTOSIS_CLOUD_CONNECT_URL}>*/}
+      {/*    <NavButton label={"Link your CLI"} Icon={<PiLinkSimpleBold />} isActive={true} />*/}
+      {/*  </Link>*/}
+      {/*)}*/}
       {/*<Link to={"/catalog"}>*/}
       {/*  <NavButton label={"View catalog"} Icon={<FiPackage />} isActive={location.pathname.startsWith("/catalog")} />*/}
       {/*</Link>*/}